move airgap pull to channel.ts
