+++ src/index.ts
@@ -1,6 +1,6 @@
 export { VendorPortalApi } from "./configuration";
 export { getApplicationDetails } from "./applications";
-export { Channel, createChannel, getChannelDetails, archiveChannel } from "./channels";
+export { Channel, createChannel, getChannelDetails, archiveChannel, pollForAirgapReleaseStatus } from "./channels";
 export { ClusterVersion, createCluster, createClusterWithLicense, pollForStatus, getKubeconfig, removeCluster, upgradeCluster, getClusterVersions, createAddonObjectStore, pollForAddonStatus, exposeClusterPort } from "./clusters";
 export { KubernetesDistribution, archiveCustomer, createCustomer, getUsedKubernetesDistributions } from "./customers";
-export { Release, CompatibilityResult, createRelease, createReleaseFromChart, promoteRelease, reportCompatibilityResult, pollForAirgapReleaseStatus } from "./releases";
+export { Release, CompatibilityResult, createRelease, createReleaseFromChart, promoteRelease, reportCompatibilityResult } from "./releases";
